Removed cotire
cotire has proven itself to be counter-productive. It rebuilts the pre-compiled
header after any header in pulseview is modified causes a complete rebuild of
all of PulseView. Also, the pre-compiled header contains a union of all the
includes of all source files, which masks errors where source files are missing
needed includes.